refactor: trip report components
